+++ script/tool/lib/src/main.dart
@@ -16,6 +16,7 @@ import 'drive_examples_command.dart';
 import 'firebase_test_lab_command.dart';
 import 'format_command.dart';
 import 'license_check_command.dart';
+import 'lint_android_command.dart';
 import 'lint_podspecs_command.dart';
 import 'list_command.dart';
 import 'native_test_command.dart';
@@ -51,6 +52,7 @@ void main(List<String> args) {
     ..addCommand(FirebaseTestLabCommand(packagesDir))
     ..addCommand(FormatCommand(packagesDir))
     ..addCommand(LicenseCheckCommand(packagesDir))
+    ..addCommand(LintAndroidCommand(packagesDir))
     ..addCommand(LintPodspecsCommand(packagesDir))
     ..addCommand(ListCommand(packagesDir))
     ..addCommand(NativeTestCommand(packagesDir))